No longer allow editing items in response to cell activate events.
The only possible ways of editing an item are by pressing the F2 key, or double-clicking over it. Previously, editing could also be started by pressing the enter, intro and space keys, or by a single-clicking over an item; we no longer allow any of those. Now they simply result in an activate() event being sent to an item. The possibility to begin an edit operation by single clicking over an item also conflicted with drag-and-drop, and that's the primary reason behind only allowing double click for editing purposes.