Improve the code that handles the selections to restore those selections lost due to an item's parent category being collapsed.
From now on, the only way Sidebar.selected can be null is by setting it explicitly to that value from client code, and thus we handle that case separately in Tree.set_selected. THIS CANNOT HAPPEN IN RESPONSE TO USER INTERACTION. For example, if an item is un-selected because its parent category has been collapsed, then it will remain as the current selected item (not in reality, just as the value of Sidebar.selected) and will be re-selected after the parent category is expanded again. THIS ALL HAPPENS SILENTLY BEHIND THE SCENES, so client code will never know it ever happened; the value of Sidebar.selected remains unchanged and Sidebar.item_selected is not emitted.