[API] Add Granite.Widgets.ThinPaned.vala
This widget does exactly the same as SidebarPaned, but its implementation is a lot different (and shorter). The overlay handle is a GdkWindow attached on top of a GtkPaned, which allows to export its API directly. This also allows using the widget either as a GtkContainer, GtkOrientable, GtkPaned or a ThinPaned, without the fear of different methods being called for each of them.

This widget doesn't suffer from the transparency issue affecting SidebarPaned, nor the implementation drawback that would render GtkWidget.parent unusable for SidebarPaned's child widgets (it would return the internal GtkPaned instead of a reference to the SidebarPaned widget).