[API] [ABI] Deprecate SidebarPaned widget in favor of ThinPaned.
The new widget is not binary compatible with the old one; nor is the C API, which means that every app using SidebarPaned will need to be re-compiled. As the ThinPaned Vala API is *equivalent* to the old SidebarPaned API, and SidebarPaned now inherits from it, this won't require making any change to the code written in Vala.