Re-write code responsible for handling level indentation.
We no longer rely on GtkTreeView.level_indentation for indentation purposes, because it had the bad side-effect of also adding extra space on the right-side, breaking the alignment of the expander and activatable-icon renderers.

The new code packs empty cell renderers at the beginning of every new row, keeping visible only those with index lower or equal than the current indentation level of the row.