ThinPaned: draw pane separator by itself rather than using the one rendered by the base class.
This fixes some theming issues, and makes sure that the sidebar-pane-separator style class is only used when rendering the separator, so that it doesn't affect the entire background area. This bug was noticeable in granite-demo and pretty much any other widget that let the background of ThinPaned pass through it.