Complete implementation of DnD reordering of items.
This is done through the SourceListSortable interface, which needs to be implemented by expandable items.

This interface is also honored by the ROOT item, which means that it's possible to re-order main categories through DnD. Although the model-side implementation and logic is completed, it doesn't work yet because category headers consume click events by expanding/collapsing. We need to find a way to keep this feature while still allowing users to drag categories around.