[Settings] Don't disconnect signals every time a key is saved: it looks like it does not always work. Use a boolean to avoid mutual signal triggering. The behavior of connecting and disconnecting at such a fast rate (i.e. without an iteration of the main loop) must be undefined.
Comments welcomed, but keep in mind that the fact that this is undefined means that the old code could perfectly work in most cases. The disconnecting/connecting can be triggered up to the number of the keys in the same iteration in normal apps. Using handler uid does not help.

I specifically encountered problem with gala notifications problem, but it is one of the only schema that is frequently modified by two different apps at the same time.