Remove path from --shared-library flag
See issue #241 for context.

Passing libgranite.full_path() to g-ir-compiler during the build
results in a typelib which references the local build path, which
doesn't correspond to the actual build path in the binary that's
shipped to the users in the packaged version of libgranite.

Adding the library as input to the custom_target() rule and referencing
it via @PLAINNAME@ only leaves its file name, without its full path, in
the invocation of g-ir-compiler.